+++ packages/core/src/slideshow/index.ts
@@ -1,3 +1,2 @@
-export * from "./slideshow.types";
-export * from "./parseSlideshow";
-export { isSceneLikeCompositionId } from "./sceneId";
+// Moved to @hyperframes/parsers/slideshow. Re-exported here for back-compat.
+export * from "@hyperframes/parsers/slideshow";
