+++ packages/core/src/utils/urlPath.ts
@@ -1,11 +1,2 @@
-export function decodeUrlPathVariants(path: string): string[] {
-  const variants = [path];
-  try {
-    const decoded = decodeURIComponent(path);
-    if (decoded !== path) variants.unshift(decoded);
-  } catch {
-    // Malformed percent sequences may be literal filesystem names.
-  }
-
-  return variants;
-}
+// Moved to @hyperframes/parsers. Re-exported here for back-compat.
+export { decodeUrlPathVariants } from "@hyperframes/parsers";
