+++ packages/parsers/src/assets.ts
@@ -1,0 +1,4 @@
+// Node-only asset-path utilities (use node:path). Behind the
+// @hyperframes/parsers/asset-paths subpath so the main entry stays browser-safe.
+export * from "./assetPaths.js";
+export * from "./rewriteSubCompPaths.js";
