+++ packages/parsers/src/index.ts
@@ -4,3 +4,15 @@ export * from "./htmlParser.js";
 export * from "./hfIds.js";
 export { unrollComputedTimeline } from "./gsapUnroll.js";
 export { queryByAttr } from "./utils/cssSelector.js";
+
+// Pure, browser-safe composition primitives shared by the linter (so it can
+// consume them without depending on @hyperframes/core). The Node-only asset
+// path helpers live behind the ./asset-paths subpath to keep this entry
+// browser-safe.
+export { decodeUrlPathVariants } from "./utils/urlPath.js";
+export {
+  FONT_ALIAS_MAP,
+  FONT_ALIAS_KEYS,
+  CANONICAL_FONT_DISPLAY_NAMES,
+  resolveAliasDisplayName,
+} from "./fontAliases.js";
