+++ packages/parsers/src/slideshow/index.ts
@@ -1,0 +1,3 @@
+export * from "./slideshow.types.js";
+export * from "./parseSlideshow.js";
+export { isSceneLikeCompositionId } from "./sceneId.js";
