+++ packages/parsers/src/slideshow/parseSlideshow.ts
