+++ packages/parsers/src/slideshow/slideshow.types.ts
