+++ packages/parsers/tsup.config.ts
@@ -10,6 +10,8 @@ export default defineConfig({
     springEase: "src/springEase.ts",
     hfIds: "src/hfIds.ts",
     gsapParser: "src/gsapParser.ts",
+    slideshow: "src/slideshow/index.ts",
+    assets: "src/assets.ts",
   },
   format: ["esm"],
   outDir: "dist",
